refactor: :art: Mejornado repository
Clientes y facturacion en repository

+++ PonchoPOS/src/main/java/com/example/PonchoPOS/repository/FacturacionRepository.java
@@ -3,5 +3,5 @@ package com.example.PonchoPOS.repository;
 import com.example.PonchoPOS.model.Facturacion;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FacturacionRepository extends JpaRepository<Facturacion, Long> {
-}
+public interface FacturacionRepository extends JpaRepository<Facturacion, Integer> {
+    }
